requirements.txt
diffusers==0.12.1
flax==0.6.4
fonttools==4.38.0
ftfy==6.1.1
huggingface-hub==0.12.0
jax==0.4.2
numpy==1.24.1
Pillow==9.4.0
python-dotenv==0.21.1
requests==2.28.2
tqdm==4.64.1
transformers==4.24.0
requests-oauthlib==1.3.1
spacy==3.5.0
openai==0.26.4
#en-core-web-sm==3.5.0
nltk==3.8.1
torch==1.13.1
torchvision==0.14.1
pandas==1.5.3
spacy-download==1.0.0